+++ ntex-bytes/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [1.10.0] (unreleased)
+
+* Add `BytePage::info()` to report the kind of storage backing a page
+
 ## [1.9.0] (2026-07-17)
 
 * Mark `BytePage::as_ptr()` as `unsafe`, as it has additional requirements that make it unsafe